+++ .travis.yml
@@ -5,10 +5,6 @@ python:
     - "3.7"
     - "3.8"
 
-# pytest-codestyle ; python_version >= "3.5"
-# pytest-mypy ; platform_python_implementation != "PyPy" and python_version >= "3.5"
-# pytest-runner
-
 install:
     - pip install pytest-codestyle
     - pip install pytest-mypy
