+++ arpeggio/arpeggio_settings.py
@@ -1,0 +1,2 @@
+
+DEFAULT_WS = '\t\n\r '
