+++ arpeggio/tests/regressions/issue_16/test_issue_16.py
@@ -5,12 +5,12 @@ from typing import Any, List
 try:
     # imports for local pytest
     from ....arpeggio import ParseTreeNode  # type: ignore # pragma: no cover
-    from ....cleanpeg import ParserPEG      # type: ignore # pragma: no cover
+    from ....parser_peg_clean import ParserPEG      # type: ignore # pragma: no cover
 except ImportError:                         # type: ignore # pragma: no cover
     # imports for doctest
     # noinspection PyUnresolvedReferences
     from arpeggio import ParseTreeNode      # type: ignore # pragma: no cover
-    from cleanpeg import ParserPEG          # type: ignore # pragma: no cover
+    from parser_peg_clean import ParserPEG          # type: ignore # pragma: no cover
 
 
 parser_input = """\
