+++ arpeggio/tests/regressions/issue_43/test_issue43.py
@@ -5,11 +5,11 @@ See https://github.com/textX/Arpeggio/issues/43
 # proj
 try:
     # imports for local pytest
-    from ....cleanpeg import ParserPEG      # type: ignore # pragma: no cover
+    from ....parser_peg_clean import ParserPEG      # type: ignore # pragma: no cover
 except ImportError:                         # type: ignore # pragma: no cover
     # imports for doctest
     # noinspection PyUnresolvedReferences
-    from cleanpeg import ParserPEG          # type: ignore # pragma: no cover
+    from parser_peg_clean import ParserPEG          # type: ignore # pragma: no cover
 
 
 def try_grammar(peg: str) -> None:
