--- perf-tests/run_all_py2.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-./run_memory.sh py2
-./run_speed.sh py2
-
