+++ perf-tests/run_speed.sh
@@ -2,4 +2,3 @@
 mkdir -p reports
 python --version > reports/${1}_speed_report.txt 2>&1
 python test_speed.py >> reports/${1}_speed_report.txt
-
