+++ perf-tests/test_memory_memoization.py
@@ -1,11 +1,9 @@
-#-*- coding: utf-8 -*-
 #######################################################################
 # Purpose: Testing memory consumption with memoization enabled.
 # Author: Igor R. Dejanovic <igor DOT dejanovic AT gmail DOT com>
 # Copyright: (c) 2016 Igor R. Dejanovic <igor DOT dejanovic AT gmail DOT com>
 # License: MIT License
 #######################################################################
-from __future__ import print_function, unicode_literals
 
 import codecs
 from os.path import dirname, join
